requirements.txt
aiofiles==23.1.0
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
fastapi==0.68.2
filelock==3.12.0
flatbuffers==23.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
Jinja2==3.0.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
onnxruntime==1.14.1
packaging==23.1
Pillow==9.5.0
protobuf==4.22.3
pydantic==1.10.7
python-multipart==0.0.6
requests==2.28.2
starlette==0.14.2
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.15.0
